% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getQdist}
\alias{getQdist}
\title{Calculates distances between all pairwise Qcolumns}
\usage{
getQdist(modest, mc.cores = getOption("mc.cores"))
}
\arguments{
\item{modest}{A file produced by \code{importData} that contains parsed data from replicate STRUCTURE runs}

\item{mc.cores}{The number of cores to use, i.e. how many processes will be spawned (at most). By default all available cores are used.}
}
\value{
An upper diagonal matrix with pairwise distances
}
\description{
Based on information produced by \code{\link{importData}}, \code{\link{getQdist}} calculates distances between all pairwise Q-columns
}
\details{
More info to come.
}
\examples{
\dontrun{
data(MODEST)
Qdist <- getQdist(modest_MCMV)}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}
}
\seealso{
\code{\link{importData}} and \code{\link{Qdist2KDAG}}
}
\keyword{STRUCTURE,}
\keyword{getQdist,}
\keyword{importData,}

