% Generated by roxygen2 (4.0.2): do not edit by hand
\name{importData}
\alias{importData}
\title{Imports STRUCTURE data}
\usage{
importData(input.data.folder)
}
\arguments{
\item{input.data.folder}{Path to f files generated by STRUCTURE}
}
\value{
a list with the following objects. \code{Qcols}: A matrix that contains all Q-columns ordered by increasing K (individuals as rows). \code{allele_counts}: A matrix that contains allelel counts for each locus in the data set (weighted according to missing data) for each Q-column in \code{Qcols}. \code{K}: K (asssumed number of populations) for each Q-column in \code{Qcols} and \code{allele_counts}. \code{NinK}: Index of each Q-column within K. \code{NinR}: Index of each Q-column within a STRUCTURE run/replicate. \code{f_file}: The filen name from which a Q-column was extracted from. \code{file_nr}: The unique number identifier for each Qcolumn extracted from \code{f_file}
}
\description{
Imports STRUCTURE data from files given a path to folder
}
\details{
Given a path to a folder that contains f_files from replicate STRUCTURE runs spanning different Ks (number of assumed populations), \code{importData} extracts all necessary information for \code{\link{MODEST}}. Each f_file must contain a single unique numeric indentifier but may also contain any other characters. ntain any othe characters.
}
\examples{
\dontrun{
modest <- importData("path/to/folder/with/f_files")}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}
}
\seealso{
\code{\link{getQdist}} and \code{\link{Qdist2KDAG}}
}
\keyword{STRUCTURE}
\keyword{importData,}

