% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{MODEST}
\alias{MODEST}
\alias{MODEST-package}
\title{MODEST}
\description{
MODEST integrates results across replicated STRUCTURE runs spanning a large range of K values and takes into account multimodality of the data.
}
\details{
More info to come
}
\author{
Petri Kemppainen, \email{petrikemppainen2@gmail.com} and Stuart Baird, \email{stuartj.e.baird@gmail.com}
}

