% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{modest_MCMV}
\alias{modest_MCMV}
\title{Example data for MODEST}
\usage{
modest_MCMV
}
\description{
Example data for MODEST using the MCMV data (see ??? for more information).
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}
}

