% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotHist}
\alias{plotHist}
\title{Plots stacked histograms posterior probabilities for beloning to a cluster taking into account multi-modality and integrating over multiple Ks.}
\usage{
plotHist(modest, KDAGplot, K, sort.by = 1)
}
\arguments{
\item{modest}{File produced by \code{\link{importData}}.}

\item{KDAGplot}{File produced by \code{\link{plotKDAG}}.}

\item{K}{K at which histogram should be drawn}

\item{sort.by}{If single integer, gives the Qcolumn by which the stacked histogram is ordered by. If a vector, gives the order of the Q-columns.}
}
\value{
Returns the order for the Qcolumns such that all other histograms can be given the same order
}
\description{
This function uses information produced by \code{\link{getQdist}}, \code{\link{Qdist2KDAG}} and \code{\link{plotKDAG}}) and plots stacked histograms of the posterior probabilities for individuals belonging to a cluster (integrated over replicated STRUCTURE runs with different Ks).
}
\details{
More info to come
}
\examples{
data(MODEST)
KDAG <- Qdist2KDAG(Qdist_MCMV, modest_MCMV$K, merge.by="Qdist")
KDAGplot <- plotKDAG(modest_MCMV, Qdist_MCMV, KDAG)
plotHist(modest_MCMV, KDAGplot, 4, sort.by=1)
}
\author{
Petri Kemppainen, \email{petrikemppainen2@gmail.com} and Stuart Baird, \email{stuartj.e.baird@gmail.com}
}
\keyword{Qdist2KDAG,}
\keyword{STRUCTURE,}
\keyword{getQdist,}
\keyword{importData,}
\keyword{plotKDAG}

