% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Qdist2KDAG}
\alias{Qdist2KDAG}
\title{Parses STRUCTURE output to a K-directed acyclig graph (KDAG)}
\usage{
Qdist2KDAG(Qdist, K, threshold = 0.05)
}
\arguments{
\item{Qdist}{A matrix of pairwise distanances between Q-columns produced by \code{\link{getQdist}}}

\item{K}{A vector of K's for each Qcolumn in \code{Qcols}}

\item{threshold}{Gives the minimum value for the distance below which Q-columns are merged, expressed as fraction of the maximum distance between any two Q-columns (clusters) in the data set (default=0.05).}
}
\value{
returns an edge list representing the KDAG
}
\description{
Producdes a KDAG from  on single linkage clustering tree based on a pairwise matrix of distances between Q-columns (produced by \code{\link{getQdist}}).
}
\details{
More info to come
}
\examples{
data(MODEST)
KDAG <- Qdist2KDAG(Qdist_MCMV, modest_MCMV$K)
}
\author{
Petri Kemppainen, \email{petrikemppainen2@gmail.com} and Stuart Baird, \email{stuartj.e.baird@gmail.com}
}
\seealso{
\code{\link{getQdist}} and \code{\link{plotKDAG}}
}
\keyword{Qdist2KDAG,}
\keyword{getQdist}

