% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Qdist_MCMV}
\alias{Qdist_MCMV}
\title{Qdist based on the G-test}
\usage{
Qdist_MCMV
}
\description{
Pairwise distances based on the G-test (\code{\link{likelihood.test}} from the \code{\link{Deducer}} package) produced by function \code{\link{getQdist}}.
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}
}

