% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getQdist}
\alias{getQdist}
\title{Estimate cluster distance (Qdist)}
\usage{
getQdist(modest, mc.cores = 1)
}
\arguments{
\item{modest}{A file produced by \code{importData} that contains parsed data from STRUCTURE output data}

\item{mc.cores}{The number of cores to use, i.e. how many processes will be spawned (at most). Default is one (1). Using more than one core in GUI or embedded environmentsis is not recommended.}
}
\value{
An upper diagonal matrix with pairwise Q-distances (\code{Qdist})
}
\description{
Based on information produced by \code{\link{importData}}, \code{\link{getQdist}} calculates cluster distances between all pairwise Q-columns
}
\details{
We use a likelihood ratio test of independence (G-test: \code{Deducer::likelihood.test}) of allele counts between clusters (contribution from each locus is weighted by the proportion of missing). Based on this distance, \code{Qdist}, a single linkage clustering tree is generated (see \code{\link{hclust}}), method="single" for details) which is then parsed into  a K-directed acyclic graph (KDAG) by the function: \code{\link{Qdist2KDAG}}. Also function \code{\link{plotKDAG}} requires \code{Qdist} as an input.
}
\examples{
\dontrun{
data(MODEST)
Qdist <- getQdist(modest_MCMV)}
}
\author{
Petri Kemppainen \email{petrikemppainen2@gmail.com}
}
\seealso{
\code{\link{importData}}, \code{\link{Qdist2KDAG}} and \code{\link{plotKDAG}}
}
\keyword{STRUCTURE,}
\keyword{getQdist,}
\keyword{importData,}

