% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotHist}
\alias{plotHist}
\title{Plots stacked histograms posterior probabilities (Q-values)}
\usage{
plotHist(modest, KDAGplot, K, sort.by = NA)
}
\arguments{
\item{modest}{File produced by \code{\link{importData}}.}

\item{KDAGplot}{File produced by \code{\link{plotKDAG}}.}

\item{K}{K at which histogram should be drawn}

\item{sort.by}{If \code{NA} (default) individuals appear in the same order as in the original data. If single integer, gives the Qcolumn by which the stacked histogram is ordered by. If a vector, gives the order of the Q-columns.}
}
\value{
Returns a list with three objects. \code{order}: the order for the Qcolumns which can be used to give all subsequent histograms the same order. \code{Qcol_index}: the index of cloums of file \code{modest$Qcols} which represented each vertex in the KDAG (this information comes from file \code{KDAGplot} produced by function \code{\link{plotKDAG}}). \code{hist.data} gives the mean Q-value for each vertex, where individuals are ordered as defined by \code{sort.by}.
}
\description{
This function uses information produced by \code{\link{getQdist}}, \code{\link{Qdist2KDAG}} and \code{\link{plotKDAG}}) and plots stacked histograms of the posterior probabilities for individuals belonging to a cluster (integrated over replicated STRUCTURE runs with different Ks).
}
\details{
More info to come
}
\examples{
data(MODEST)
KDAG <- Qdist2KDAG(Qdist_MCMV, modest_MCMV$K, merge.by="Qdist")
KDAGplot <- plotKDAG(modest_MCMV, Qdist_MCMV, KDAG)
out <- plotHist(modest_MCMV, KDAGplot, 2, sort.by=1)
names(out)
out <- plotHist(modest_MCMV, KDAGplot, 4, sort.by=out$order)
}
\author{
Petri Kemppainen, \email{petrikemppainen2@gmail.com} and Stuart Baird, \email{stuartj.e.baird@gmail.com}
}
\keyword{Qdist2KDAG,}
\keyword{STRUCTURE,}
\keyword{getQdist,}
\keyword{importData,}
\keyword{plotKDAG}

